+++ src/presentation/http/errors.ts
@@ -1,7 +1,26 @@
 import { CodedError } from '@/shared/coded-error';
 
-export class BadRequest extends CodedError {
+export class HttpError extends CodedError {
+  statusCode: number;
+  constructor(
+    code: string,
+    message: string,
+    statusCode: number,
+    details?: Record<string, any>
+  ) {
+    super(code, message, details);
+    this.statusCode = statusCode;
+  }
+}
+
+export class BadRequest extends HttpError {
   constructor(code: string, message: string, details?: Record<string, any>) {
-    super(code, message, details, 400);
+    super(code, message, 400, details);
+  }
+}
+
+export class NotFoundError extends HttpError {
+  constructor() {
+    super('NOT_FOUND', 'Page not found', 404);
   }
 }