+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Ongoing Changes
 
+- Added route status
+- Added Decorator version
 - Added redis module
 - Added http integration folder
 - Added test unit for useCases